python/openai_react_agent/pyproject.toml
[project]
name = "openai-react-agent"
version = "0.1.0"
description = "Async OpenAI React Agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "acontext>=0.0.4",
    "openai>=1.51.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]


